+++ Dockerfile
@@ -10,11 +10,6 @@ RUN apt-get update && apt-get install -y \
 # Set up working directory
 WORKDIR /app
 
-# Copy package files and install Node.js dependencies
-COPY package.json pnpm-lock.yaml ./
-RUN npm install -g pnpm && pnpm install
-
-# Install Python dependencies globally (override the system protection)
 COPY requirements.txt .
 RUN pip3 install --break-system-packages --no-cache-dir -r requirements.txt
 
@@ -25,4 +20,4 @@ COPY . .
 EXPOSE 8000
 
 # Start both services with correct working directories
-CMD ["sh", "-c", "cd BE && python manage.py migrate && cd .. && pnpm dev:be"]
+CMD ["sh", "-c", "cd BE && python manage.py migrate && cd ../FE/web-ui && pnpm dev:be"]
