+++ FE/web-ui/.gitignore
@@ -9,10 +9,10 @@
 !.yarn/plugins
 !.yarn/releases
 !.yarn/versions
-
+pnpm-lock.yaml
 # testing
 /coverage
-
+/reports
 # next.js
 /.next/
 /out/
