+++ FE/web-ui/jest.config.ts
@@ -1,8 +1,10 @@
 import type { Config } from "jest";
 import nextJest from "next/jest.js";
 
+// 1. Update this to your new relative path. 
+// If your Next.js files are now in the root, use "./"
 const createJestConfig = nextJest({
-  dir: "./FE/web-ui",
+  dir: "./", 
 });
 
 const config: Config = {
@@ -10,7 +12,8 @@ const config: Config = {
   testEnvironment: "jsdom",
   setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
   collectCoverage: true,
-  collectCoverageFrom: ["FE/web-ui/src/app/**/*.{ts,tsx,js,jsx}"],
+  // 2. Update these globs to reflect your new folder structure
+  collectCoverageFrom: ["src/app/**/*.{ts,tsx,js,jsx}"], 
   coverageThreshold: {
     global: {
       branches: 70,
@@ -18,7 +21,8 @@ const config: Config = {
       lines: 70,
       statements: 70,
     },
-    "FE/web-ui/src/app/**/*.ts?(x)": {
+    // 3. Update the specific path threshold
+    "src/app/**/*.ts?(x)": {
       branches: 70,
       functions: 70,
       lines: 70,
@@ -28,11 +32,9 @@ const config: Config = {
   coverageDirectory: "coverage",
   coverageReporters: ["json", "text", "lcov"],
   clearMocks: true,
-  transform: {
-    "^.+\\.(ts|tsx)$": "ts-jest",
-  },
   moduleNameMapper: {
-    "^@/(.*)$": "<rootDir>/$1",
+    // This ensures your @/ aliases still work from the new root
+    "^@/(.*)$": "<rootDir>/src/$1", 
   },
   testMatch: ["**/*.spec.ts", "**/*.spec.tsx"],
 };
