+++ FE/web-ui/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "scripts": {
     "dev": "concurrently -k -n FRONT,BACK -c blue,green \"pnpm dev:fe\" \"pnpm dev:be\"",
-    "dev:fe": "next dev FE/web-ui",
-    "dev:be": "cd BE && python manage.py runserver 0.0.0.0:8000",
-    "build": "next build FE/web-ui",
+    "dev:fe": "next dev",
+    "dev:be": "cd ../../ && docker compose start",
+    "build": "next build",
     "build:tokens": "cd ./FE/web-ui/ && style-dictionary build --config ./style-dictionary.config.js --verbose   # build .scss styles from .json tokens",
-    "start": "next start FE/web-ui",
-    "lint": "next lint FE/web-ui",
-    "eslint --fix": "eslint FE/web-ui --fix",
+    "start": "next start",
+    "lint": "next lint",
+    "eslint --fix": "eslint --fix",
     "test:unit": "jest --coverage",
     "test:e2e": "(cucumber-js e2e/features  || true) && tsx e2e/helper/report/report.ts"
   },
