+++ FE/web-ui/tsconfig.json
@@ -22,8 +22,15 @@
     "paths": {
       "@/*": ["./src/*"]
     },
-    "types": ["@testing-library/jest-dom"],
+    "types": ["@testing-library/jest-dom"]
   },
-  "include": ["next-env.d.ts", ".next/types/**/*.ts", "**/*.ts", "**/*.tsx", "../../eslint.config.ts", "style-dictionary.config.js"],
+  "include": [
+    "next-env.d.ts",
+    ".next/types/**/*.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    "eslint.config.ts",
+    "style-dictionary.config.js"
+  ],
   "exclude": ["node_modules"]
 }
